deployment/requirements.txt
tensorflow==2.15.0
streamlit==1.44.0
seaborn==0.12.2
pandas==1.5.3
matplotlib==3.7.1
plotly==5.15.0
pillow==9.5.0
numpy==1.24.3